Making radial chart clean
